% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tgconfig.R
\name{dump_example_config}
\alias{dump_example_config}
\title{Dump example config file}
\usage{
dump_example_config(path)
}
\arguments{
\item{path}{path to dump example config file to}
}
\description{
Dump example config file
}
