% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tgconfig.R
\name{get_param_strict}
\alias{get_param_strict}
\title{Get package parameters and return error if they do not exist}
\usage{
get_param_strict(param, package = NULL)
}
\arguments{
\item{param}{parameter to get}

\item{package}{package of the parameter (NULL if running from a package)}
}
\value{
value of \code{param} in package \code{package} and error if parameter no found
}
\description{
Get package parameters and return error if they do not exist
}
\examples{

register_param('param1', 'tgconfig')
set_param('param1', 'value1', 'tgconfig')
get_param_strict('param1', 'tgconfig')

# try to get a parameter that doesn't exist
# get_param_strict('other_param', 'tgconfig') # would throw an error


}
