% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tgconfig.R
\name{guess_package}
\alias{guess_package}
\title{Guess package the code is invoked from}
\usage{
guess_package(env)
}
\arguments{
\item{env}{environment}
}
\value{
name of the package the code is invoked from, and NULL if it is not run from a package
}
\description{
Guess package the code is invoked from
}
