% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tgconfig.R
\name{has_param}
\alias{has_param}
\title{Check if package has a paramter}
\usage{
has_param(param, package = NULL)
}
\arguments{
\item{param}{parameter}

\item{package}{package}
}
\description{
Check if package has a paramter
}
