% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tgconfig.R
\name{list_package_params}
\alias{list_package_params}
\title{List package parameters}
\usage{
list_package_params(package)
}
\arguments{
\item{package}{package}
}
\value{
names of package parameters. NULL if \code{package} has no parameters
}
\description{
List package parameters
}
