% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tgconfig.R
\name{load_params_to_env}
\alias{load_params_to_env}
\title{Load parameters to current environment}
\usage{
load_params_to_env(params, package = NULL, envir = parent.frame())
}
\arguments{
\item{params}{parameters to load}

\item{package}{package}

\item{envir}{environment to load to}
}
\value{
invisibly returns the changed environment
}
\description{
Load paramters as variables to the current environment (or any other environment \code{envir})
}
\examples{
register_params(example_config_file(), 'tgconfig')
get_package_params('tgconfig')
load_params_to_env(c('expr_param', 'boolean_param'), 'tgconfig')
expr_param
boolean_param

}
