% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tgconfig.R
\name{override_params}
\alias{override_params}
\title{Override pre-set parameters from config file}
\usage{
override_params(config_file, package = NULL)
}
\arguments{
\item{config_file}{yaml file with parameters and values}

\item{package}{package}
}
\description{
Override pre-set parameters from config file
}
\examples{
config_file <- example_config_file()
register_params(config_file, 'tgconfig')
get_package_params('tgconfig')
override_params(system.file('config/override_example.yaml', package='tgconfig'), package='tgconfig')
get_package_params('tgconfig')


}
