% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tgconfig.R
\name{rm_package_params}
\alias{rm_package_params}
\title{Remove all package parameters}
\usage{
rm_package_params(package = NULL)
}
\arguments{
\item{package}{package}
}
\description{
Remove all package parameters
}
\examples{
config_file <- example_config_file()
register_params(config_file, 'tgconfig')
get_package_params('tgconfig')
rm_package_params('tgconfig')
get_package_params('tgconfig')
}
