% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tgconfig.R
\name{rm_param}
\alias{rm_param}
\title{Remove parameter}
\usage{
rm_param(param, package = NULL)
}
\arguments{
\item{param}{parameter to remove}

\item{package}{package}
}
\description{
Remove parameter
}
\examples{
register_param('param1', 'tgconfig')
has_param('param1', 'tgconfig')
rm_param('param1', 'tgconfig')
has_param('param1', 'tgconfig')
}
