% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tgconfig.R
\name{set_param}
\alias{set_param}
\title{Set package parameter}
\usage{
set_param(param, value, package = NULL)
}
\arguments{
\item{param}{parameter to set. An error would be thrown if parameter is not registered.}

\item{value}{value to set the parameter to}

\item{package}{package of the parameter (NULL if running from a package)}
}
\description{
Set package parameter
}
\examples{
register_param('param1', 'tgconfig')
set_param('param1', 'value1', 'tgconfig')
get_param('param1', 'tgconfig')

# try to set a parameter that doesn't exist
# set_param('other_param', 'tgconfig') # would thorw an error

}
\seealso{
register_param, get_param
}
