% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tgconfig.R
\name{get_package_params}
\alias{get_package_params}
\title{Get all package parameters}
\usage{
get_package_params(package)
}
\arguments{
\item{package}{package}
}
\value{
a list with package parameters and values. NULL if \code{package} has no parameters
}
\description{
Get all package parameters
}
