% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tgconfig.R
\name{override_params}
\alias{override_params}
\title{Override pre-set parameters from config file}
\usage{
override_params(config_file, package = NULL)
}
\arguments{
\item{config_file}{yaml file with paramters and values}

\item{package}{package}
}
\description{
Override pre-set parameters from config file
}
