% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tgconfig.R
\name{register_param}
\alias{register_param}
\title{Register a parameter to package}
\usage{
register_param(param, package = NULL, default_value = NA)
}
\arguments{
\item{param}{parameter to register}

\item{package}{package to register parameter to}

\item{default_value}{default value of the parameter (default: NA)}
}
\description{
Register a parameter to package
}
\examples{
register_param('param1', 'tgconfig')
get_package_params('tgconfig')

}
