% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tgconfig.R
\name{register_params}
\alias{register_params}
\title{Register parameters from config file}
\usage{
register_params(config_file, package = NULL)
}
\arguments{
\item{config_file}{yaml file with paramters and values}

\item{package}{package}
}
\description{
Register parameters from config file
}
